+++ index.js
@@ -8,5 +8,6 @@ const Record = require('./src/source').Record
 const Result = require('./src/job').Result
 const Source = require('./src/source').Source
 const Store = require('./src/store')
+const Helper = require('./src/helper')
 
-module.exports = { BatchRunner, Config, Interceptor, Job, Printer, Record, Result, Source, Store }
+module.exports = { BatchRunner, Config, Interceptor, Job, Printer, Record, Result, Source, Store, Helper }